+++ src/lib/merchant_api_history.c
@@ -163,8 +163,8 @@ struct TALER_MERCHANT_HistoryOperation *
 TALER_MERCHANT_history (struct GNUNET_CURL_Context *ctx,
                         const char *backend_url,
                         const char *instance,
-                        unsigned int start,
-                        unsigned int delta,
+                        unsigned long long start,
+                        long long delta,
                         struct GNUNET_TIME_Absolute date,
                         TALER_MERCHANT_HistoryOperationCallback history_cb,
                         void *history_cb_cls)
@@ -181,7 +181,7 @@ TALER_MERCHANT_history (struct GNUNET_CURL_Context *ctx,
   seconds = date.abs_value_us / 1000LL / 1000LL;
   base = TALER_url_join (backend_url, "/history", NULL);
   GNUNET_asprintf (&ho->url,
-                   "%s?date=%llu&instance=%s&start=%d&delta=%d",
+                   "%s?date=%llu&instance=%s&start=%llu&delta=%lld",
                    base,
                    seconds,
                    instance,
